requirements.txt
asgiref==3.5.2
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
ConfigArgParse==1.5.3
coverage==7.0.3
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.14.2
MarkupSafe==2.1.1
model-bakery==1.9.0
msgpack==1.0.4
psutil==5.9.4
psycopg==3.1.3
psycopg-binary==3.1.3
psycopg2-binary==2.9.4
PyJWT==2.6.0
pytz==2022.5
pyzmq==24.0.1
requests==2.28.1
roundrobin==0.0.4
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2

